Fixed an incorrect cast.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@2985 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_array.c
@@ -1116,7 +1116,7 @@ mx_reallocate_array( void *array_pointer,
 		return NULL;
 	}
 
-	header = (unsigned long *) array_pointer;
+	header = (uint32_t *) array_pointer;
 
 	header_magic = header[MX_ARRAY_OFFSET_MAGIC];
